refactor(sqlite): separate statement planning from lexer
