+++ src/infra/adapters/sqlite/sqlite3/parser/mod.rs
@@ -1,19 +1,22 @@
 mod command_tag;
 mod lexer;
 mod output;
+mod statement_plan;
 
 pub(in crate::adapters::sqlite) use command_tag::{
     aggregate_sqlite_command_tag, command_tag_result, sqlite_statement_tags,
     statement_counts_as_select_tag,
 };
 pub(in crate::adapters::sqlite) use lexer::{
-    SqliteStatementPlan, append_changes_query_for_plan, is_create_view_prefix,
-    is_create_virtual_table_prefix, is_sqlite_rerunnable_export_query, next_keyword_from,
-    reject_sqlite_fsdir, skip_bracket_quoted, skip_quoted, sqlite_adhoc_execution_query_for_plan,
-    sqlite_empty_result_sentinel, sqlite_export_not_rerunnable_error, sqlite_probe_marker,
-    sqlite_statement_plan, virtual_table_module_name,
+    is_create_view_prefix, is_create_virtual_table_prefix, is_sqlite_rerunnable_export_query,
+    next_keyword_from, reject_sqlite_fsdir, skip_bracket_quoted, skip_quoted,
+    sqlite_export_not_rerunnable_error, virtual_table_module_name,
 };
 pub(in crate::adapters::sqlite) use output::{
     last_sqlite_result_set, parse_affected_rows, parse_count_result, quoted_to_query_result,
     strip_sqlite_probes,
+};
+pub(in crate::adapters::sqlite) use statement_plan::{
+    SqliteStatementPlan, append_changes_query_for_plan, sqlite_adhoc_execution_query_for_plan,
+    sqlite_empty_result_sentinel, sqlite_probe_marker, sqlite_statement_plan,
 };
